% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{FIC CSAT Data}
\format{
A data frame with 6 variables:

\describe{
  \item{ID.date}{First day of each month}
  \item{NetEasy}{NetEasy rating}
  \item{Q1}{NPS rating}
  \item{Overall_why}{Free text comment; why NPS rating was given}
  \item{Q3}{Time taken to complete, compared to previous methods}
  \item{Q4}{Free text comment; suggestions for improvement of system}
}
}
\source{
{Customer Insight Team at NHS BSA, via collection FIC_data in FIC
mongoDB database}
}
\usage{
data
}
\description{
FIC CSAT Data
}
\keyword{datasets}
