% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{eval_lines}
\alias{eval_lines}
\title{Evaluate each line of plain text file}
\usage{
eval_lines(filepath)
}
\arguments{
\item{filepath}{Filepath as a String.}
}
\value{
Nothing
}
\description{
Reads a plain text file line by line, evaluating each line. Useful for
creating variables dynamically, e.g. reading in parameters.
}
\examples{
\dontrun{
filepath <- tempfile()
writeLines(
  text = "LEFT = \"right\"",
  con = filepath
)
eval_lines(filepath)
print(LEFT)
unlink(filepath) # delete temporary file
rm(left) # remove example variable
}
}
