% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{get_mongo_collection}
\alias{get_mongo_collection}
\title{Retrieve data from MongoDB}
\usage{
get_mongo_collection(collection, db, connection_string, replace_with = NULL)
}
\arguments{
\item{collection}{Name of the collection.}

\item{db}{Name of the database the collection is in.}

\item{connection_string}{Address of the mongodb server in mongo connection
string [URI
format](https://docs.mongodb.com/manual/reference/connection-string/).}

\item{replace_with}{Default is NULL. If passed, must be a dataframe, named
list (for single record) or character vector with json strings (one string
for each row).}
}
\value{
Returns all data in the given collection as a dataframe.
}
\description{
Retrieve all data from a mongodb collection. Optionally,
  replace the collection data first.
}
\examples{
\dontrun{
# Get the data in my_collection (in the my_db database)
get_mongo_collection(
  "my_collection", "my_db",
  connection_string = glue::glue(
    "mongodb+srv://{MONGO_USER}:{MONGO_PASS}@{MONGO_HOST}",
    "/?tls=true&retryWrites=true&w=majority"
  )
)

# Replace the data in my_collection (in the my_db database) with the data in
# my_dataframe, then get that data
get_mongo_collection(
  "my_collection", "my_db",
  connection_string = glue::glue(
    "mongodb+srv://{MONGO_USER}:{MONGO_PASS}@{MONGO_HOST}",
    "/?tls=true&retryWrites=true&w=majority"
  ),
  replace_with = my_dataframe
)
}
}
