% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{load_data}
\alias{load_data}
\title{Load data from mongoDB}
\usage{
load_data(r, data)
}
\arguments{
\item{r}{A reactiveValues object}

\item{data}{Either "data" or "targets"}
}
\value{
Nothing - a new child of r is created to hold the data.
}
\description{
Attempts to read data from specified mongoDB collection. Once
read, the raw data is cleaned and prepared. If an error occurs in either of
these 2 steps, data is read from an .rda file in the data folder - this
contains the last successfully read data from the mongoDB collection.
}
\examples{
\dontrun{
r <- reactiveValues()

load_data(r, "data")
load_data(r, "targets")
}
}
