% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\docType{data}
\name{targets}
\alias{targets}
\title{FIC CSAT Targets}
\format{
A data frame with 3 variables:
\describe{
  \item{ID.date}{First day of each month}
  \item{NPS.Target}{Target Net Promoter Score}
  \item{NES.Target}{Target Net Easy Score}
}
}
\source{
{Customer Insight Team at NHS BSA, via collection FIC_targets in
FIC mongoDB database}
}
\usage{
targets
}
\description{
FIC CSAT targets.
}
\keyword{datasets}
